How to Use the Project
The platform works in simple and direct way, users should go through the list of items and select the ones they are willing to purchase. 

Once you're done selecting, click the cart to check your list of items, you can also remove items in this category.

Proceed to checkout page where you'll make payment as required. Note: the payment mode is in a testing mode.